+++ src/main/java/com/natalia/proyectoSpringBoot/controllers/CourseController.java
@@ -2,7 +2,7 @@ package com.natalia.proyectoSpringBoot.controllers;
 
 import com.natalia.proyectoSpringBoot.dto.CourseDTO;
 import com.natalia.proyectoSpringBoot.models.Course;
-import com.natalia.proyectoSpringBoot.services.CourseServiceImpl;
+import com.natalia.proyectoSpringBoot.services.CourseService;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
 
@@ -10,36 +10,37 @@ import java.util.List;
 import java.util.Map;
 
 @RestController
+@RequestMapping("/courses")
 public class CourseController {
-    private final CourseServiceImpl courseServiceImpl;
+    private final CourseService courseService;
 
-    public CourseController(CourseServiceImpl courseServiceImpl){
-        this.courseServiceImpl = courseServiceImpl;
+    public CourseController(CourseService courseService){
+        this.courseService = courseService;
     }
 
-    @GetMapping("/courses")
+    @GetMapping()
     public List<Course> getCourses(){
-        return this.courseServiceImpl.getCourses();
+        return this.courseService.getCourses();
     }
 
-    @GetMapping("/courses/info")
+    @GetMapping("/{info}")
     public List <CourseDTO> getCoursesInfo(){
-        return this.courseServiceImpl.getCoursesInfo();
+        return this.courseService.getCoursesInfo();
     }
 
-    @PostMapping(value = "/courses", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @PostMapping(value = {""}, consumes = MediaType.APPLICATION_JSON_VALUE)
     public Course createCourse(@RequestBody Map<String, String> json){
-        return this.courseServiceImpl.add(json.get("name"), json.get("career"));
+        return this.courseService.add(json.get("name"), json.get("career"));
     }
 
-    @PutMapping(value = "/courses/{idCourse}", consumes = MediaType.APPLICATION_JSON_VALUE)
+    @PutMapping(value = "/{idCourse}", consumes = MediaType.APPLICATION_JSON_VALUE)
     public Course updateCourse(@RequestBody Course course, @PathVariable Long idCourse){
-            return this.courseServiceImpl.updateById(course, idCourse);
+            return this.courseService.updateById(course, idCourse);
     }
 
-    @DeleteMapping(value = "/courses/{idCourse}")
+    @DeleteMapping(value = "/{idCourse}")
     public void deleteCourse(@PathVariable Long idCourse){
-            this.courseServiceImpl.deleteById(idCourse);
+            this.courseService.deleteById(idCourse);
 
     }
